no-mistakes(review): extract season helper to import-light module; add paddleocr-free tests
